templated particle loop with access descriptors
