+++ include/neso_particles.hpp
@@ -9,10 +9,12 @@
 #include "cell_binning.hpp"
 #include "cell_dat.hpp"
 #include "compute_target.hpp"
+#include "containers/tuple.hpp"
 #include "domain.hpp"
 #include "global_mapping.hpp"
 #include "local_mapping.hpp"
 #include "local_move.hpp"
+#include "loop/particle_loop.hpp"
 #include "mesh_hierarchy.hpp"
 #include "mesh_interface.hpp"
 #include "mesh_interface_local_decomp.hpp"
